requirements.txt
opencv-python-headless>=4.8.0
numpy>=1.24.0
pytesseract>=0.3.10
flask>=3.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
pdf2image>=1.16.0
